requirements.txt
Flask
Flask-Session
redis
pandas
openpyxl
python-dotenv